cria biblioteca <vector.h> de interacoes entre os pixels

+++ ppm.h
@@ -5,7 +5,7 @@
 #include <stdlib.h>
 
 typedef struct {
-	float red, green, blue;
+	double red, green, blue;
 } PPMPixel;
 
 typedef struct {
